Adds urls to iframe in lti search
