+++ src/containers/ErrorPage/__tests__/ErrorPage-test.js
@@ -9,9 +9,9 @@
 
 import React from 'react';
 import renderer from 'react-test-renderer';
+import { MissingRouterContext } from '@ndla/safelink';
 import serializer from 'jest-emotion';
 import IntlProvider from '@ndla/i18n';
-import { BrowserRouter } from 'react-router-dom';
 import ErrorPage from '../ErrorPage';
 import { getLocaleObject } from '../../../i18n';
 
@@ -25,9 +25,9 @@ test('ErrorPage renderers correctly', () => {
   const locale = getLocaleObject('nb');
   const component = renderer.create(
     <IntlProvider locale={locale.abbreviation} messages={locale.messages}>
-      <BrowserRouter>
+      <MissingRouterContext.Provider value={true}>
         <ErrorPage locale="nb" location={{ pathname: '/' }} />
-      </BrowserRouter>
+      </MissingRouterContext.Provider>
     </IntlProvider>,
   );
