+++ src/containers/SearchPage/SearchContainer.jsx
@@ -145,6 +145,7 @@ class SearchContainer extends Component {
       ltiData,
       enabledTabs,
       allTabValue,
+      isLti,
     } = this.props;
     const { subjects } = data;
     const { query } = this.state;
@@ -221,7 +222,7 @@ class SearchContainer extends Component {
                 locale,
               )
             : [];
-
+          console.log(search);
           return (
             <SearchPage
               closeUrl="/#"
@@ -250,6 +251,7 @@ class SearchContainer extends Component {
                 onUpdateContextFilters={this.onUpdateContextFilters}
                 includeEmbedButton={includeEmbedButton}
                 ltiData={ltiData}
+                isLti={isLti}
               />
               {isReadyToShow && (
                 <Pager
@@ -298,6 +300,7 @@ SearchContainer.propTypes = {
     }),
   ),
   allTabValue: string,
+  isLti: bool,
 };
 
 SearchContainer.defaultProps = {
@@ -307,6 +310,7 @@ SearchContainer.defaultProps = {
   data: {},
   handleSearchParamsChange: () => {},
   allTabValue: 'all',
+  isLti: false,
 };
 
 export default injectT(SearchContainer);
