+++ src/containers/SearchPage/searchHelpers.js
@@ -5,6 +5,7 @@ import { contentTypeIcons } from '../../constants';
 import { getContentType } from '../../util/getContentType';
 import LtiEmbed from '../../lti/LtiEmbed';
 import { toSubjects } from '../../routeHelpers';
+import { parseAndMatchUrl } from '../../util/urlHelper';
 
 const getRelevance = resource => {
   if (resource.filters.length > 0) {
@@ -136,17 +137,36 @@ export const convertResult = (results, subjectFilters, enabledTab, language) =>
     };
   });
 
+const getResultUrl = (id, url, isLti) => {
+  if ((url && url.href) || !isLti) {
+    return url;
+  }
+
+  const parsedUrl = parseAndMatchUrl(url);
+  if (!parsedUrl) {
+    return url;
+  }
+  return `/article-iframe/nb/urn:resource:${parsedUrl.params.resourceId}/${id}`;
+};
 export const resultsWithContentTypeBadgeAndImage = (
   results,
   t,
   includeEmbedButton,
   ltiData,
+  isLti,
 ) =>
   results.map(result => {
     const { contentType } = result;
-
+    console.log(result);
     return {
       ...result,
+      url: getResultUrl(result.id, result.url, isLti),
+      urls: isLti
+        ? result.urls.map(urlObject => ({
+            ...urlObject,
+            url: getResultUrl(result.id, urlObject.url, isLti),
+          }))
+        : result.urls,
       contentType,
       contentTypeIcon: contentTypeIcons[contentType || result.contentType] || (
         <ContentTypeBadge
