+++ src/server/routes/oembedArticleRoute.js
@@ -6,41 +6,13 @@
  *
  */
 
-import { matchPath } from 'react-router-dom';
 import { BAD_REQUEST, INTERNAL_SERVER_ERROR } from 'http-status';
-import parseUrl from 'parse-url';
-import { isValidLocale } from '../../i18n';
 import { getArticleIdFromResource } from '../../containers/Resources/resourceHelpers';
 import { fetchResource } from '../../containers/Resources/resourceApi';
-import { RESOURCE_PAGE_PATH, PLAIN_ARTICLE_PAGE_PATH } from '../../constants';
 import config from '../../config';
 import handleError from '../../util/handleError';
 import { fetchArticle } from '../../containers/ArticlePage/articleApi';
-
-export function matchUrl(pathname, isPlainArticle, lang = false) {
-  if (isPlainArticle) {
-    return matchPath(
-      pathname,
-      lang ? `/:lang${PLAIN_ARTICLE_PAGE_PATH}` : PLAIN_ARTICLE_PAGE_PATH,
-    );
-  }
-  return matchPath(
-    pathname,
-    lang ? `/:lang${RESOURCE_PAGE_PATH}` : RESOURCE_PAGE_PATH,
-  );
-}
-
-export function parseAndMatchUrl(url) {
-  const { pathname } = parseUrl(url);
-  const paths = pathname.split('/');
-  paths[1] = paths[1] === 'unknown' ? 'nb' : paths[1];
-  const path = paths.join('/');
-
-  if (isValidLocale(paths[1])) {
-    return matchUrl(path, paths[2] === 'article', true);
-  }
-  return matchUrl(path, paths[1] === 'article', false);
-}
+import { parseAndMatchUrl } from '../../util/urlHelper';
 
 function getOembedObject(req, title, html) {
   return {
@@ -93,8 +65,8 @@ export async function oembedArticleRoute(req) {
 
     return getOembedObject(
       req,
-      resource.title,
-      `<iframe aria-label="${resource.title}" src="${config.ndlaFrontendDomain}/article-iframe/${lang}/${resource.id}/${articleId}?removeRelatedContent=true" frameborder="0" allowFullscreen="" />`,
+      resource.name,
+      `<iframe aria-label="${resource.name}" src="${config.ndlaFrontendDomain}/article-iframe/${lang}/${resource.id}/${articleId}?removeRelatedContent=true" frameborder="0" allowFullscreen="" />`,
     );
   } catch (error) {
     handleError(error);
